优化竞技场查询代码
